move all code under odyssey namespace
Noticed that we're adding `Formula` into the top-level namespace, which is pretty
bad practice, so I went ahead and moved all the formula code under `lib/odyssey`,
and wrapped it in the odyssey module, like a typical gem.
