+++ lib/odyssey/formulas/coleman_liau.rb
@@ -1,4 +1,4 @@
-class Ari < Formula
+class ColemanLiau < Odyssey::Formula
 
   def score(text, stats)
     calc_score(stats['letter_count'], stats['word_count'], stats['sentence_count'])
@@ -11,14 +11,13 @@ class Ari < Formula
                           stats_split['word_count'][i],
                           1))
     end
-    res
   end
 
   def calc_score(letter_count, word_count, sentence_count)
-    (((4.71 * (letter_count.to_f / word_count.to_f)) + (0.5 * (word_count.to_f / sentence_count.to_f))) - 21.43).round(1)
+    ((5.89 * (letter_count.to_f / word_count.to_f)) - (0.3 * (sentence_count.to_f / word_count.to_f)) - 15.8).round(1)
   end
 
   def name
-    'Automated Readability Index'
+    'Coleman-Liau Index'
   end
 end
